% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParam.R
\name{derivedParam}
\alias{derivedParam}
\title{Compute a derived parameter.}
\usage{
derivedParam(.self, param, dparamname, constants)
}
\arguments{
\item{.self}{An \code{MLEModel} object.}

\item{param}{A vector of parameters at which to calculate the derived parameters.}

\item{dparamname}{The name of the derived parameter.}

\item{constants}{A list of constants used to calculate the derived parameters.}
}
\description{
This method takes a vector of parameters, constants and the name of a derived
parameter, and returns the value of the derived parameter. In \code{MLEModel}, this method simply
returns an empty value. Implementing subclasses typically populates this method to handle all
derived parameters listed in the \code{dparamlist} field of the model using \code{switch} protocol.
}
