% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawErrors.R
\name{drawErrors}
\alias{drawErrors}
\title{Draws error vectors for an \code{MLEModel} object.}
\usage{
drawErrors(.self, data, simopts)
}
\arguments{
\item{.self}{An \code{MLEModel} object.}

\item{data}{An \code{MLEData} object.}

\item{simopts}{An \code{MLESimulationOptions} object.}
}
\description{
Draws error vectors for an \code{MLEModel} object.
}
