% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawUnobservables.R
\name{drawUnobservables}
\alias{drawUnobservables}
\title{Draws unobservable vectors for an \code{MLEModel} object}
\usage{
drawUnobservables(.self, data, simopts)
}
\arguments{
\item{.self}{An \code{MLEModel} object.}

\item{data}{An \code{MLEData} object.}

\item{simopts}{An \code{MLESimulationOptions} object.}
}
\description{
Draws unobservable vectors for an \code{MLEModel} object
}
