% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate a Maximum Likelihood Estimation Model}
\usage{
estimate(.self, data, estopts = NULL)
}
\arguments{
\item{.self}{An \code{MLEModel} object.}

\item{data}{An \code{MLEData} object.}

\item{estopt}{An \code{MLEEstimationOptions} object.}
}
\description{
Estimate a Maximum Likelihood Estimation Model
}
