% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expandArray}
\alias{expandArray}
\title{Replicate each row of a matrix a specified number of times}
\usage{
expandArray(array, counts)
}
\arguments{
\item{array}{A vector or a matrix whose rows will be replicated.}

\item{counts}{A vector or a single-column matrix with length equal to the number of rows in
\code{array}. The \code{i}th row of \code{array} will be replicated \code{counts[i]} times.}
}
\value{
A matrix of which the number of rows is the summation of all elements in \code{counts}.
}
\description{
Replicate each row of a matrix a specified number of times
}
