% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDerivedParam.R
\name{getDerivedParam}
\alias{getDerivedParam}
\title{Compute all derived parameters of the model.}
\usage{
getDerivedParam(.self, param, constants = NULL)
}
\arguments{
\item{.self}{An \code{MLEModel} object.}

\item{param}{A vector of parameters at which to calculate the derived parameters.}

\item{constants}{A list of constants used to to calculate the derived parameters.}
}
\description{
This method is a wrapper of the method \code{derivedParam}, which returns a vector
of all derived parameters.
}
