% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{groups}
\alias{groups}
\title{Create a vector indexing the groups defined by the columns of a matrix}
\usage{
groups(x)
}
\arguments{
\item{x}{A vector or a matrix that defines groups.}
}
\value{
A vector of group indices.
}
\description{
Create a vector indexing the groups defined by the columns of a matrix
}
