% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEData.R
\docType{class}
\name{MLEData-class}
\alias{MLEData-class}
\alias{MLEData}
\title{A Reference Class that Defines Data Used in \code{MLEModel}.}
\description{
A Reference Class that Defines Data Used in \code{MLEModel}.
}

