% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEEstimationOptions.R
\docType{class}
\name{MLEEstimationOptions-class}
\alias{MLEEstimationOptions-class}
\alias{MLEEstimationOptions}
\title{A Reference Class that Defines Options for the \code{estimate} Method of \code{MLEModel} Class.}
\description{
This class is based on the \code{optim} optimization method.
}
\section{Fields}{

\describe{
\item{\code{startparam}}{Starting paramter vector.}

\item{\code{method}}{The optimization method to be used. Default is L-BFGS-B.}

\item{\code{maxit}}{The maximum number of iterations. Default is 10,000.}

\item{\code{constr}}{An \code{MLEConstraints} object.}

\item{\code{hesstol}}{Numerical step for computing the Hessian.}

\item{\code{quadacc}}{Accuracy of quadrature nodes for numerical integration in estimation.}

\item{\code{quadacc_deriv}}{Accuracy of quadrature nodes for numerical integration in computing Hessians, gradients, etc.}

\item{\code{first_step_vcov}}{Variance-Covariance matrix for first estimation step}

\item{\code{first_step_paramlist}}{A list of parameter names from first estimation step.}

\item{\code{first_step_param}}{A vector with the values of the first step parameters.}

\item{\code{compute_hessian}}{Compute Hessian at the estimated parameters.}

\item{\code{compute_jacobian}}{Compute Jacobian of likelihood vector at the estimated parameters.}
}}


