% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEEstimationOutput.R
\docType{class}
\name{MLEEstimationOutput-class}
\alias{MLEEstimationOutput-class}
\alias{MLEEstimationOutput}
\title{A Reference Class that Defines Options for the \code{simulate} Method of \code{MLEModel} Class.}
\description{
A Reference Class that Defines Options for the \code{simulate} Method of \code{MLEModel} Class.
}
\section{Fields}{

\describe{
\item{\code{constr}}{An \code{MLEConstaints} object.}

\item{\code{dparam}}{A vector of derived parameters.}

\item{\code{const}}{A list of constants.}
}}


