% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLESetOfDatasets.R
\docType{class}
\name{MLESetOfDatasets-class}
\alias{MLESetOfDatasets-class}
\alias{MLESetOfDatasets}
\title{A Reference Class that Holds a Vector of \code{MLEData} Objects.}
\description{
A Reference Class that Holds a Vector of \code{MLEData} Objects.
}
\section{Fields}{

\describe{
\item{\code{datasets}}{A list of MLEData objects}

\item{\code{ndatasets}}{The number of MLEData objects}
}}

\section{Methods}{

\describe{
\item{\code{addDataset(dataset)}}{Add a new MLEData object in the list of datasets.

         \code{dataset}: An MLEData object to be added.}

\item{\code{loadDatasetsFromDisk(directory, name, indices, collapseArrayVars = 1)}}{Loops over stored data and create an \code{MLESetOfDatasets} object according to
         the \code{loadFromDisk} method of \code{MLEData}.

         \code{directory}: The location of files to be loaded.
         \code{name}: The name of files to be saved.
         \code{precision}: The number of decimal to store for all inputs. Default is 8.
         \code{indices}: Positional indices of the datasets to save.
         \code{collapseArrayVars}: Whether array variables are collapsed. Default is collapse.}

\item{\code{saveDatasetsToDisk(directory, name, precision = 8,
  indices = 1:.self$ndatasets)}}{Loops over and store data in an \code{MleSetOfDatasets} object according to
         the \code{saveToDisk} method of \code{MLEData}.

         \code{directory}: The location of files to be saved.
         \code{name}: The name of files to be saved.
         \code{precision}: The number of decimal to store for all inputs. Default is 8.
         \code{indices}: Positional indices of the datasets to save.}
}}

