% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEConstraints.R
\docType{class}
\name{MLEConstraints-class}
\alias{MLEConstraints-class}
\alias{MLEConstraints}
\title{A Reference Class that Defines Constraints of Parameters for a Maximum Likelihood Model.}
\description{
A Reference Class that Defines Constraints of Parameters for a Maximum Likelihood Model.
}
\section{Fields}{

\describe{
\item{\code{lower}}{The lower limit of the parameters.}

\item{\code{upper}}{The upper limit of the parameters.}
}}


