% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEEstimationOutput.R
\docType{class}
\name{MLEEstimationOutput-class}
\alias{MLEEstimationOutput-class}
\alias{MLEEstimationOutput}
\title{A Reference Class that Holds Estimates of the Maximum Likelihood Estimation model}
\description{
A Reference Class that Holds Estimates of the Maximum Likelihood Estimation model
}
\section{Fields}{

\describe{
\item{\code{hessian}}{The numerical Hessian matrix evaluated at estimated parameters.}

\item{\code{jacobian}}{The numerical Jacobian matrix evaluated at estimated parameters.}

\item{\code{constr}}{An \code{MLEConstaints} object.}

\item{\code{dparam}}{A vector of derived parameters.}

\item{\code{const}}{A list of constants used in estimation.}
}}


\section{Fields}{


\describe{
\item{\code{param}}{Locally optimal solution.}

\item{\code{fval}}{The value of objective function at the optimal solution.}

\item{\code{convergence}}{An integer indicating convergence of estimation. 0 indicates successful completion.}

\item{\code{message}}{A character string giving any additional information returned by the optimizer.}

\item{\code{estopts}}{A \code{ModelEstimationOptions} object.}

\item{\code{model}}{A \code{Model} object used for estimation.}

\item{\code{nobs}}{The number of observations in the data used for estimation.}

\item{\code{vcov}}{Variance-covariance matrix of estimated parameters.}

\item{\code{se}}{Standard errors of estimated parameters.}
}
}

