% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEModel.R
\docType{class}
\name{MLEModel-class}
\alias{MLEModel-class}
\alias{MLEModel}
\title{A Reference Class that Provides a Template for Maximum Likelihood Models}
\description{
A Reference Class that Provides a Template for Maximum Likelihood Models
}
\section{Fields}{

\describe{
\item{\code{group_unobs_list}}{A vector of names of group-level unobservables (integrated numerically).}

\item{\code{indiv_unobs_list}}{A vector of names of individual-level unobservables (integrated numerically).}

\item{\code{error_list:}}{A vector of names of individual-level errors (not integrated numerically).}

\item{\code{error_distributions}}{A list that contains the distributions of all errors in \code{error_list}.}

\item{\code{error_dimensions}}{A list that contains the number of dimensions of each observations of
all errors in \code{error_list}.}

\item{\code{dparamlist}}{A vector of names of derived parameters.}

\item{\code{ngroup_unobs}}{The number of group-level unobservables in \code{group_unobs_list}.}

\item{\code{nindiv_unobs}}{The number of individual-level unobservables in \code{indiv_unobs_list}.}

\item{\code{nerrors}}{The number of error terms in \code{error_list}.}

\item{\code{ndparam}}{The number of derived parameters in \code{dparamlist}.}

\item{\code{dindices}}{A list that gives the index of each derived parameter.}

\item{\code{numerical_integral}}{An indicator of whether the model requires numerical integration.}
}}


