% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLESetOfDatasets.R
\docType{class}
\name{MLESetOfDatasets-class}
\alias{MLESetOfDatasets-class}
\alias{MLESetOfDatasets}
\title{A Reference Class that Holds a List of Datasets}
\description{
A Reference Class that Holds a List of Datasets
}
\section{Fields}{

\describe{
\item{\code{datasets}}{A list of \code{MLEData} objects.}

\item{\code{ndatasets}}{The number of \code{MLEData} objects.}
}}

\section{Methods}{

\describe{
\item{\code{addDataset(dataset)}}{\subsection{Description}{
        Add a new \code{MLEData} object in the list of datasets.}

        \subsection{Parameters}{
        \code{dataset}: An \code{MLEData} object to be added.}}

\item{\code{loadDatasetsFromDisk(directory, name, indices, collapseArrayVars = 1)}}{\subsection{Description}{
        Loops over saved datasets and create an \code{MLESetOfDatasets} object according to
        the \code{loadFromDisk} method of \code{MLEData}.}

        \subsection{Parameters}{
        \code{directory}: The location where files will be loaded. 

        \code{name}: The name of the files to be loaded.

        \code{indices}: The positional indices of the datasets to load.

        \code{collapseArrayVars}: Whether array variables are collapsed. Default is collapse.}}

\item{\code{saveDatasetsToDisk(directory, name, precision = 8, indices = NULL)}}{\subsection{Description}{
        Loops over and save datasets in an \code{MLESetOfDatasets} object according to
        the \code{saveToDisk} method of \code{MLEData}.}

        \subsection{Parameters}{
        \code{directory}: The location where files will be saved. 

        \code{name}: The name of the files to be saved.

        \code{precision}: The number of decimal places to save. Default is 8.

        \code{indices}: The positional indices of the datasets to save. If not specified,
        all datasets wil be saved.
}}
}}

