% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLESetOfEstimates.R
\docType{class}
\name{MLESetOfEstimates-class}
\alias{MLESetOfEstimates-class}
\alias{MLESetOfEstimates}
\title{A Reference Class that Holds a List of Estimation Results}
\description{
A Reference Class that Holds a List of Estimation Results
}
\section{Fields}{

\describe{
\item{\code{estimates}}{A list of \code{MLEEstimationOutput} objects.}

\item{\code{nest}}{The number of estimates.}

\item{\code{param_list}}{A list of parameter estimates.}

\item{\code{dparam_list}}{A list of of derived parameter estimates.}

\item{\code{startparam_list}}{A list of starting parameters.}

\item{\code{dstartparam_list}}{A list of derived parameters implied by starting parameters.}
}}


